Tweaked parse_to_table
Removed the "we want only lines that we can split" condition since
we might want to define rows that have only the first column in
a multi-column table (because we're lazy). Extra (empty) rows are
easily removed by clicking the corresponding [x] button anyway.

Also removed code that attaches the change_delimiter function to
the delimiter input field's onchange attribute; attached that
function to the appropriate button instead.